move checkScopes to getScopes array
