+++ apps/xyz/mod/workspace/getScopes.js
@@ -117,3 +117,59 @@ export function scopesTree(scopes) {
 
   return tree;
 }
+
+/**
+@function checkScope
+
+@description
+The checkScope method checks whether the user has access to the object based on the provided roles and templateScope.
+
+The method is exported so that hosts which resolve roles outside the workspace API can evaluate scope strings with the same semantics applied during composition.
+
+If the roles parameter is falsy, the method will return false.
+
+If the roles parameter is true, the method will return true.
+
+If the roles parameter is an array, the method will check whether the templateScope string is included in the roles array. If so, it will return true.
+
+@param {array} templateScope
+@param {array} roles
+@returns {boolean} Returns true if the user has access based on the roles and templateScope.
+*/
+export function checkScope(templateScope, roles) {
+  // The templateScope array is empty, meaning there are no access restrictions.
+  if (!templateScope.length) return true;
+
+  // Prevent access if no roles are provided from user.
+  if (!roles) return false;
+
+  // Admin endpoints will set the roles parameter to true to bypass role checks.
+  if (roles === true) return true;
+
+  // Filter out undefined values from the templateScope array and join the remaining values with a pipe character to create a string representation of the template scope.
+  const templateScopeString = templateScope.join('.');
+
+  // Check whether the roles array includes the templateScopeString.
+  if (roles.includes(templateScopeString)) {
+    return true;
+  }
+
+  // Access should be granted if the templateScopeString is the first part of any nested role.
+  if (
+    roles.findIndex((role) => role.startsWith(`${templateScopeString}.`)) > -1
+  ) {
+    return true;
+  }
+
+  if (templateScopeString === '*') return true;
+
+  // Validate access if roles array contains every scope in the templateScope array. Must be enabled in xyzEnv.LEGACY_ROLES to allow for legacy role checks. This is a temporary solution to allow for legacy role checks until the roles are refactored to be more granular and hierarchical.
+  if (
+    xyzEnv.LEGACY_ROLES &&
+    templateScope.every((scope) => roles.includes(scope))
+  ) {
+    return true;
+  }
+
+  return false;
+}
